Publish online or offline message to status topic
Changes to set a will message that is published to the status topic to indicate the esp module is offline.  Also, publish a message to the same topic to indicate "online" when MQTT connects successfully.  This is useful for sending notifications (such as push notification via pushbullet) to the user when the esp module goes offline.